Allow for smaller values of a (1e-7 and below)
This makes the test added in b48a42f3fee407f4133c1457dc3945be565f1c75
(as failing) pass.

The problem is this: The interface C code that is called from within R
code passes parameters to the actual Triangle library almost the same
way as a command line invocation would: in the form of a string (on the
command line it would be a list of strings). This string consists of
substrings of the form “q30”, “a0.01”, etc. just like during a command
line invocation one would pass “-q30” and “-a0.01”, the difference being
the missing leading dash and a lack of separation between the
substrings (the final string could look like this: “q30a0.01").
Consequently, the way this string is parsed is in the form “read
a/q/etc. and then as many digits or periods as there are”.

The string “a0.01” is created by an invocation of sprintf with the
format string “a%f” (and is later read back in using strtod). The number
of digits after the period is left unspecified here, causing the default
of 6 to be used. Consequently, numbers that are (rounded to) values of
1e-7 and below (this includes e.g. 4e-7) are printed as 0 and
consequently received by the Triangle code as 0, causing an error with
the claim that maxarea needs to be positive and is not.

Fixing this problem is unfortunately not straightforward at all. A few
options come to mind:

* Instead of a string, one could pass a struct to the triangle code.
  This should be faster and cleaner than the current solution,
  eliminating all problems of this type. It would be rather invasive,
  though, and undesirable in the sense that changes to triangle.c should
  be kept to a minimum to lower the maintenance burden.
* One could replace the format string %f with %g or %G, so that 1e-7
  could be printed as “1e-7” or “1E-7". Unfortunately, both e and E are
  also accepted as switches by the triangle code, so that ‘a1.0e5’ and
  ‘a1.0e’ would have to be handled completely differently.
* Finally (and that is what this commit does), one could go with %.Nf
  where N is a number that is computed on the fly from the argument: If
  I pass 1e-7, I need 7 digits after the period, with 1e-10 I’d need 10
  and so forth (it seems to be a good idea to use a few more digits
  actually so that with 1e-7, 1.1e-7, and 1.000001e-7 can still be
  distinguished).